Added check for inserted calculator into makemove

+++ makemove.py
@@ -3,6 +3,9 @@ import shutil
 
 
 os.system("make")
-shutil.copyfile("fx-CG50TETRIS.g3a", "E:/fx-CG50TETRIS.g3a")
 
-print("Move was successful")
+if os.path.exists("E:"):
+    shutil.copyfile("fx-CG50TETRIS.g3a", "E:/fx-CG50TETRIS.g3a")
+    print("Move was successful")
+else:
+    print("Drive not inserted, move not attempted")
